simplified avoidance
As robot will travel with field map, we have detected blocks mared on
the map, robot doesn't necessarily to avoid the block but go somethere
else. We can also optimize the calculate the path of grids to travel to
avoid the block on the path instead of solely rely on avoidacne class.